requirements.txt
streamlit
pandas
google-cloud-bigquery
google-auth
google-cloud-bigquery-storage